Split `core::alloc` module
